GW3S contracts: added getTokens for levelLoupeFacet

+++ contracts/GW3S/libraries/LibAppStorage.sol
@@ -16,7 +16,7 @@ struct AppStorage {
     mapping(address => mapping(uint256 => address)) level_instance;      // The current instance of player deployed by the a specific level 
     // type = enum{LEVEL, HIDDEN, HACKER}
     mapping(uint => mapping(uint => address)) Erc721LevelReward;         // Store ERC721 smart contract by (levelId => type => svgContract)
-    mapping(uint256 => mapping(uint => address)) level_factories;                          // The address of factory by level
+    mapping(uint256 => mapping(uint8 => address)) level_factories;        // The address of factory by level
     mapping(uint256 => address[]) level_tokens;                          // The array of tokens by level
 }
